requirements.txt
# Core Data Science & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
scipy>=1.11.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# FastAPI & Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Experiment Tracking & MLOps
wandb>=0.15.0
scikit-optimize>=0.9.0

# Monitoring & Metrics
prometheus-client>=0.17.0

# Frontend
streamlit>=1.28.0
requests>=2.31.0

# Testing & Code Quality
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.27.0
flake8>=6.1.0
pylint>=3.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
loguru>=0.7.0

# Development Tools
black>=23.10.0
isort>=5.12.0
